+++ src/check.rs
@@ -208,7 +208,7 @@ impl VirtualEnv {
                 self.handle_args_outputs(1, parts.len().saturating_sub(1))
             }
             Node::Unpack { count, .. } => self.handle_args_outputs(1, *count),
-            Node::Mod(Astar, args, _) | Node::ImplMod(AstarFirst, args, _) => {
+            Node::ImplMod(Astar, args, _) | Node::ImplMod(AstarFirst, args, _) => {
                 self.pop();
                 let [neighbors, heuristic, is_goal] = get_args(args)?;
                 let has_costs = neighbors.outputs() == 2;